+++ pymongo/mongo_client.py
@@ -704,7 +704,6 @@ class MongoClient(common.BaseObject):
         self.__kill_cursors_queue = []
 
         self._event_listeners = options.pool_options.event_listeners
-
         super(MongoClient, self).__init__(options.codec_options,
                                           options.read_preference,
                                           options.write_concern,
@@ -1127,10 +1126,10 @@ class MongoClient(common.BaseObject):
         sending one or more endSessions commands.
 
         Close all sockets in the connection pools and stop the monitor threads.
-        If this instance is used again it will be automatically re-opened and
-        the threads restarted unless auto encryption is enabled. A client
-        enabled with auto encryption cannot be used again after being closed;
-        any attempt will raise :exc:`~.errors.InvalidOperation`.
+
+        .. versionchanged:: 4.0
+           Once closed, the client cannot be used again and any attempt will
+           raise :exc:`~pymongo.errors.InvalidOperation`.
 
         .. versionchanged:: 3.6
            End all server sessions created by this client.